Update workflow
This commit improves the workflow for this chart by:

1. Using joinaggregate to calculate the total amount of loss per class, total amount of loss per year, and max total loss in a specific year

Using joinaggregate allows us to more easily filter out classes without any loss and also reduces the number of tables needed in the VEGA. The previous workflow used aggregate which resulted in lots of tables. It also filtered out all rows with 0 loss which resulted in losing data points for years with no loss.